# apps/cms/openapi.yaml
openapi: 3.1.0
info:
  title: LumiBase CMS API
  version: 0.1.0
  description: |
    Edge-native Headless CMS API. See `docs/api/hono-api-spec.md` for the full
    surface; this file is the machine-readable contract that ships with the
    repo. Endpoints land incrementally — Phase 0 covers auth/health only.
servers:
  - url: https://api.lumibase.dev/api/v1
    description: Production
  - url: http://127.0.0.1:1989/api/v1
    description: Local wrangler dev
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    SiteHeader:
      name: X-Lumi-Site
      in: header
      required: true
      schema: { type: string }
      description: Active tenant id. Required for every authenticated endpoint.
  schemas:
    Error:
      type: object
      properties:
        errors:
          type: array
          items:
            type: object
            required: [code, message]
            properties:
              code: { type: string }
              message: { type: string }
              path:
                type: array
                items: { type: string }
              trace: { type: object, additionalProperties: true }
...
    FlowGraph:
      type: object
      properties:
        entry: { type: string, nullable: true }
        nodes:
          type: array
          items:
            type: object
            required: [id, key]
            properties:
              id: { type: string }
              key: { type: string }
              options: { type: object, additionalProperties: true }
              next: { type: string, nullable: true }
              onError: { type: string, nullable: true }
    FlowResource:
      type: object
      required: [name, triggerType]
      properties:
        id: { type: string }
        siteId: { type: string }
        name: { type: string }
        description: { type: string, nullable: true }
        status: { type: string, enum: [active, inactive, draft], default: draft }
        triggerType: { type: string, enum: [webhook, event, schedule, manual] }
        triggerOptions: { type: object, additionalProperties: true }
        graph: { $ref: '#/components/schemas/FlowGraph' }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    FlowRun:
      type: object
      properties:
        id: { type: string }
        flowId: { type: string }
        siteId: { type: string }
        status: { type: string, enum: [running, success, error] }
        input: { type: object, additionalProperties: true }
        steps:
          type: array
          items: { type: object, additionalProperties: true }
        error: { type: string, nullable: true }
        startedAt: { type: string, format: date-time }
        finishedAt: { type: string, format: date-time, nullable: true }
    FlowRunResult:
      type: object
      properties:
        runId: { type: string }
        status: { type: string, enum: [success, error] }
        steps:
          type: array
          items: { type: object, additionalProperties: true }
        error: { type: string, nullable: true }
    # --- Marketplace (POST-GA3) ---
...
paths:
...
  /flows:
    get:
      summary: List flows in the active site
      parameters: [{ $ref: '#/components/parameters/SiteHeader' }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/FlowResource' }
    post:
      summary: Create a flow
      parameters: [{ $ref: '#/components/parameters/SiteHeader' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FlowResource' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/FlowResource' }
        '400':
          description: Validation error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /flows/{id}:
    parameters:
      - $ref: '#/components/parameters/SiteHeader'
      - { name: id, in: path, required: true, schema: { type: string } }
    get:
      summary: Get flow detail (incl. graph)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/FlowResource' }
        '404':
          description: Not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    patch:
      summary: Update flow (graph, status, metadata)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FlowResource' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/FlowResource' }
    delete:
      summary: Delete a flow
      responses:
        '204': { description: Deleted }
  /flows/{id}/run:
    parameters:
      - $ref: '#/components/parameters/SiteHeader'
      - { name: id, in: path, required: true, schema: { type: string } }
    post:
      summary: Trigger a manual flow run
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              description: Optional input data passed to the flow entry node
              additionalProperties: true
      responses:
        '200':
          description: Run result
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/FlowRunResult' }
  /flows/{id}/runs:
    parameters:
      - $ref: '#/components/parameters/SiteHeader'
      - { name: id, in: path, required: true, schema: { type: string } }
    get:
      summary: List runs for a flow (latest 100)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/FlowRun' }

  # ---------------------------------------------------------------------------
  # Marketplace (POST-GA3)
  # ---------------------------------------------------------------------------
...